chore: refactor to create base_gan_model
